refactor: remove QuizEditor component
Deleted the QuizEditor.tsx file, which appears to be a previous implementation of the quiz module editor. This removal likely indicates a significant refactoring or replacement of the quiz editing functionality.
